feat(vscode): add isolated extension launch config
Adds a new launch configuration "Run Extension [Isolated]" to `.vscode/launch.json`. This configuration allows running the VS Code extension in an isolated environment, disabling other extensions and pointing to a specific launch root.

I find myself changing this in a lot of my branches so I thought it might be nice to make it an official configuration!

A `launch/README.md` file is also added to serve as a placeholder for the isolated launch root.